+++ internal/machine/spectrum/sound/ay38912_test.go
@@ -1,0 +1,10 @@
+package sound
+
+import (
+	"mcs/testutils/dsl"
+	"testing"
+)
+
+func TestAY38912(t *testing.T) {
+	dsl.RunScenarios(t, ay38912Scenarios)
+}