+++ internal/machine/spectrum/sound/main_test.go
@@ -1,0 +1,11 @@
+package sound
+
+import (
+	"testing"
+
+	"mcs/testutils"
+)
+
+func TestMain(m *testing.M) {
+	testutils.RunWithAssertTracking(m)
+}